requirements.txt
ably==2.0.7
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
bidict==0.23.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.15
distro==1.9.0
faiss-cpu==1.9.0.post1
Flask==3.0.3
flask-cors==6.0.1
Flask-Login==0.6.3
fonttools==4.59.0
fpdf==1.7.2
frozenlist==1.5.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
kiwisolver==1.4.8
locust==2.37.14
locust-cloud==1.26.3
MarkupSafe==3.0.2
matplotlib==3.10.3
methoddispatch==3.0.2
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.1
openai==1.58.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
promptlayer==1.0.32
propcache==0.2.1
psutil==7.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pyee==11.1.1
Pygments==2.19.2
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.6
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.1
zope.interface==7.2
